feat: implement form preview component and enhance logo upload functionality
