+++ src/app/features/dashboard/dashboard.component.scss
@@ -58,7 +58,7 @@ mat-toolbar {
 @media (max-width: 768px) {
     .feedback-button {
         right: 1px;
-        bottom: 35px;
+        bottom: 75px;
     }
 }
