+++ src/app/features/forms/components/form-edit/form-preview/form-preview.component.html
@@ -1,0 +1,7 @@
+<iframe
+  id="formPreview"
+  frameborder="0"
+  class="iframe-preview"
+  [srcdoc]="iframeContent || '<p>No content available</p>'"
+  title="Form Preview">
+</iframe>
