+++ src/app/features/forms/components/form-edit/form-preview/form-preview.component.spec.ts
@@ -1,0 +1,21 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { FormPreviewComponent } from './form-preview.component';
+
+describe('FormPreviewComponent', () => {
+  let component: FormPreviewComponent;
+  let fixture: ComponentFixture<FormPreviewComponent>;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [FormPreviewComponent]
+    });
+    fixture = TestBed.createComponent(FormPreviewComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
